backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-multipart>=0.0.9
pydantic>=2.8.0
pydantic-settings>=2.4.0
openai>=1.0.0
PyYAML>=6.0.2
beautifulsoup4>=4.12.3
python-docx>=1.1.2
pypdf>=5.0.0
